requirements.txt
openai
langchain
flask_socketio
pinecone-client
tiktoken